fix(cssnano): Don't use unsafe options
Since the (S)CSS in this project is module based (css-modules), we can't use the unsafe options postcss-discard-unused and  optimise z-index and we proably should not use postcss-merge-idents too.

postcss-discard-unused is 'evil' here because for example it removes @font-face statements if it does not find any entry in the same file that uses the imported font.

But since we're using css-modules, we will proably use @font-face to import a Font, for example Roboto in our core scss file (core.scss), and then use this font in our components. But cssnano would remove the @font-face entry in this case, since there would be no use of it in core.scss.

For more information check http://cssnano.co/optimisations/